build_utils: factor out generate_c_bindings()
