+++ src/redisearch_rs/build_utils/Cargo.toml
@@ -7,6 +7,7 @@ publish.workspace = true
 
 [dependencies]
 cbindgen.workspace = true
+bindgen.workspace = true
 
 [lints]
 workspace = true
