+++ src/redisearch_rs/inverted_index_bencher/build.rs
@@ -7,10 +7,7 @@
  * GNU Affero General Public License v3 (AGPLv3).
 */
 
-use std::env;
-use std::path::PathBuf;
-
-use build_utils::{git_root, link_static_libraries, rerun_if_c_changes};
+use build_utils::{generate_c_bindings, git_root, link_static_libraries};
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     // Always link the static libraries, independent of bindgen
@@ -20,46 +17,12 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     ]);
 
     // Generate C bindings - fail build if this doesn't work
-    generate_c_bindings()?;
-
-    Ok(())
-}
-
-fn generate_c_bindings() -> Result<(), Box<dyn std::error::Error>> {
-    let root = git_root().expect("Could not find git root for static library linking");
-
-    let includes = [
-        root.join("deps").join("RedisModulesSDK"),
-        root.join("src"),
-        root.join("deps"),
-        root.join("src").join("redisearch_rs").join("headers"),
-        root.join("deps").join("VectorSimilarity").join("src"),
-        root.join("src").join("buffer"),
-    ];
-
-    let mut bindings = bindgen::Builder::default().header(
-        root.join("src")
-            .join("inverted_index")
-            .join("inverted_index.h")
-            .to_str()
-            .ok_or("Invalid path")?,
-    );
-
-    for include in includes {
-        bindings = bindings.clang_arg(format!("-I{}", include.display()));
-        // Re-run the build script if any of the C files in the included
-        // directory changes
-        let _ = rerun_if_c_changes(&include);
-    }
-
-    let out_dir = PathBuf::from(env::var("OUT_DIR")?);
-    bindings
-        .allowlist_file(".*/inverted_index.h")
-        // Don't generate the Rust exported types else we'll have a compiler issue about the wrong
-        // type being used
-        .blocklist_file(".*/types_rs.h")
-        .generate()?
-        .write_to_file(out_dir.join("bindings.rs"))?;
+    let root = git_root().expect("Could not find git root");
+    let ii_header = root
+        .join("src")
+        .join("inverted_index")
+        .join("inverted_index.h");
+    generate_c_bindings(vec![ii_header], ".*/inverted_index.h")?;
 
     Ok(())
 }
